Implemented delegate-based entity-updates for performance.
Lol, it actually works as I imagined. @main-- wanna profile?

This is a huge change, and actually I really hope that this works because
this would have some cool features. Main advantage: Not setting 1000s
of value in a slow-ass dictionary.

I've taken away support for the follwing features:
- Attribute weapons
- Bomb events

If you want fair comparison, profile master with
"DemoParser.AttributeWeapons" set to false. I really hope this is as fast
as I imagine!