feat(rank): save the rank information to avoid long loading time
The rank information for each account added to the user's list is now stored into the ranks.json file, it will check first into it and if no rank is found, it will check into the demos.
The only case that can take some times is when you have a lot of demos and a new account is added, it will search the rank from the demos that has been analyzed.
After it, it's much more faster for future analyze.
Each time a demo is anlalyzed it will check if the demo's date is more recent than the known last date, if the demo is new it will update the rank.

fixes #49 #54 #57