feat(export): export data for a specific player
This commit add the ability to focus on a specific player for excel export.
If an account is selected or if a player is selected from the demo details view, the export will focus on this player.
A popup is displayed before exporting to remind the user that he is going to export for a specific player.

fixes #58