feat(suspects): add suspects bot
The bot is a small external program that check in background if new suspects has been banned and send Windows notifications.
Since it's a independent app, it can be run independently of CSGO DM.
However the bot and the app are able to communicate, there is a "bridge"
between them that allow to start / stop the app or send messages.

Massive changes have been made to ease the bot integration.

The app (CSGO DM) now consists of 4 projects.
- Core: Core components used by the manager and the bot.
- Resources: Resources used by the manager.
- Services: Services logic used by the bot and the manager.
- Manager: The app (WPF UI)

Others changes made with this commit:
- Selected account stats are not mapped to models anymore, services must
  be used to update data
- Only selected account demos are displayed
- Some analyzes logic have been moved to collection handlers
- Suspects list is now ordered by DSLB (Ascending)
- Only one app instance launchable (manager and bot)
- Selected player stats (on demo's details view) is not saved in settings anymore
- App icon is now displayed in title bar
- Design time are now working for all views (but some doesn't have data)
- More fields have been added to json
- Some refactors in differents parts of the app

ref #97