feat(xlsx): reduce RAM usage during Excel exports
Tested with 600 demos.
Before it used ~4GB, now only ~80MB.

- Replace the Excel NPOI lib used to generate XLSX files with a custom XLSX generator.
The lib NPOI hold all data cells in memory to be able to edit them later (formatting, styling...), we don't need that.
That's why there was a huge memory spike when writing the XLSX (shave ~3GB).
It's based on the current V3 implementation.
- Hold only necessary rows data in memory instead of the whole demos (shave ~1GB)

fix #199
fix #553