feat(video): allow overlapping sequences
Now that both games (CS2/CSGO) read a JSON file to control demo's
playback, it was easier to "tell" the game to execute a different set of
commands to execute during the playback.

The JSON file now contains an array of "Sequence".
A "Sequence" contains a list of "Action" (it could have more properties
in the future).
If a sequence's actions contain the new internal command
"go_to_next_sequence", the plugin will restart the demo playback and
execute instructions of the next sequence.

ref #953